LayoutTests/inspector/styles/styles-url-linkify.html

<html>
<head>
<link rel="stylesheet" href="resources/styles-url-linkify.css">

<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/elements-test.js"></script>
<script>

function test()
{
    function completeURL(baseURL, href)
    {
        InspectorTest.addResult(WebInspector.completeURL(baseURL, href));
    }

    InspectorTest.addResult("URLs completed:");
    completeURL("http://example.com", "/");
    completeURL("http://example.com", "moo");
    completeURL("http://example.com/", "https://secure.com/moo");
    completeURL("https://example.com/foo", "//secure.com/moo");
    completeURL("http://example.com/foo/zoo", "/moo");
    completeURL("http://example.com/foo/zoo/", "moo");
    completeURL("http://example.com/foo/zoo", "boo/moo");
    completeURL("http://example.com/foo", "moo");
    completeURL("http://example.com/foo", "?a=b");
    completeURL("http://example.com/foo?c=d", "?a=b");

    const dataURL = "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAEIAAABCAgMAAACeOuh7AAAABGdBTUEAAK/INwWK6QAAAAlQTFRF////AAAA////fu+PTwAAAAF0Uk5TAEDm2GYAAACHSURBVDjLxdLbDYAgDAVQGELn0R3oEHYf2KGdUqtE46OFRCP3oyTng1xCnWsaD5JRRtCkQ2YmkBkHRXqWJBn0j0TICbrsWVoWhRShCdcGyZCtHxMaUnVPRZ9KSbmBJdsX2vJVnwqRD0Rb4rpzgIbE/AI5NTnWAMvy5l0dXrfuLh5OCe5BmmYGXhTUxlQ5xJ8AAAAASUVORK5CYII=";
    completeURL("https://example.com/foo", dataURL);

    function dumpHref()
    {
        var href;
        var valueChildNodes = WebInspector.panels.elements.sidebarPanes.styles.sections[0][2].propertiesTreeOutline.children[0].valueElement.childNodes;
        for (var i = 0; i < valueChildNodes.length; ++i) {
            if (valueChildNodes[i].href) {
                href = valueChildNodes[i].href;
                break;
            }
        }
        if (!href) {
            InspectorTest.addResult("href property not found");
            return;
        }
        var segments = href.split("/");
        var output = [];
        for (var i = segments.length - 1, minSegment = i - 3; i >= 0 && i >= minSegment; --i)
            output.unshift(segments[i]);
        InspectorTest.addResult(output.join("/"));
    }

    InspectorTest.selectNodeWithId("local");
    InspectorTest.addSniffer(WebInspector.StylesSidebarPane.prototype, "_rebuildUpdate", step1);

    function step1()
    {
        InspectorTest.addResult("Link for a URI from CSS document:");
        dumpHref();
        InspectorTest.selectNodeWithId("iframed");
        InspectorTest.addSniffer(WebInspector.StylesSidebarPane.prototype, "_rebuildUpdate", step2);

    }

    function step2()
    {
        InspectorTest.addResult("Link for a URI from iframe inline stylesheet:");
        dumpHref();
        InspectorTest.completeTest();
    }
}

</script>
</head>
<body onload="runAfterIframeIsLoaded()">
<p>
Tests that URLs are linked to and completed correctly. Bugs <a href="http://bugs.webkit.org/show_bug.cgi?id=51663">51663</a>, <a href="http://bugs.webkit.org/show_bug.cgi?id=53171">53171</a>, <a href="http://bugs.webkit.org/show_bug.cgi?id=62643">62643</a>
</p>
<div id="local"></div>
<iframe src="resources/styles-url-linkify-iframe.html"></iframe>

</body>
</html>
